+++ src/services/client.js
@@ -1,8 +1,5 @@
-import {
-    downloadContracts,
-    writeMetrics2CSV,
-    writeMetrics2JSON,
-} from "./index.js";
+import { writeMetrics2CSV, writeMetrics2JSON } from "./handle-metrics.js";
+import { downloadContracts } from "./handle-contracts.js";
 
 downloadContracts();
 // writeMetrics2CSV();
