--- src/services/index.js
@@ -1,4 +1,0 @@
-import { downloadContracts } from "./handle-contract.js";
-import { writeMetrics2CSV, writeMetrics2JSON } from "./handle-metrics.js";
-
-export { downloadContracts, writeMetrics2CSV, writeMetrics2JSON };
