--- src/utils/index.js
@@ -1,17 +1,0 @@
-import {
-    getFnFromAddr,
-    getDirFromAddr,
-    getDestFromAddr,
-    getSourceCode,
-    getUrlFromAddr,
-} from "./handle-contract-data.js";
-import getContracts from "./getContracts.js";
-
-export {
-    getContracts,
-    getFnFromAddr,
-    getDirFromAddr,
-    getDestFromAddr,
-    getSourceCode,
-    getUrlFromAddr,
-};
